Manifest functionality.
Reads manifest.xml, expands wildcards then adds all the files in the
manifest that have Archive="true" and computes hashes for them while
adding them to the archive. Finally it adds the manifest.xml file which
has now been patched to contain only the files in the archive with their
hash values.